show both tables
